Revert "Zmiany"
This reverts commit 74574dad6ead4c3bcef7b0a6e85a0b2b56e153a9.

+++ Hikari Engine/src/Window.cpp
@@ -12,8 +12,8 @@ void Hikari::Window::position(int x, int y)
 
 void Hikari::Window::position(Hikari::Vector2D& point)
 {
-	m_PosX = static_cast<unsigned int>(point.x());
-	m_PosY = static_cast<unsigned int>(point.y());
+	m_PosX = point.x();
+	m_PosY = point.y();
 }
 
 Hikari::Vector2D Hikari::Window::position()
